Use pixishape to define the hit area
When a terrain is not controlled, using the pixishape to define the hit area allows users to more easily draw adjacent terrain without inadvertently clicking the existing terrain. Particularly for polygons. 

Some bug, maybe in the PIXI.Ellipse contains method, is causing the hit area to be the center of the ellipse. Selecting ellipses by drawing a small rectangle around the center works, but the hit area for ellipses is otherwise incorrect. Could either fix the Ellipse contains method or fall back on the boundary rectangle.